Fix MID-4465: extension delta serialization
For deltas referencing c:ExtensionType or its concrete substitutes
(like {...}UserExtensionType) the serializer produced xsi:type
attribute that caused deserialization to fail.

The solution is

1. to forbid producing explicitly typed XNodes if the type is runtime
(in prism serializer, bean serializer and in RawType serialization);

2. to change semantics of definition.isRuntimeSchema - now it is
true if and only if the definition itself is part of the runtime schema.
Not to be confused with isXsdAnyMarker and/or isDynamic!